fixed key readings

+++ src/main/java/pl/grizwold/chip8/Chip8EmulatorStarter.java
@@ -89,7 +89,7 @@ public class Chip8EmulatorStarter extends Application {
             short keyboard = 0;
             for (int i = 0; i < keys.length; i++) {
                 if (keys[i]) {
-                    keyboard = (short) ((short) (1 << i) & 0xFFFF);
+                    keyboard = (short) ((1 << i) & 0xFFFF);
                 }
             }
             return keyboard;
